Added a TransactionBuilder, added TransactionManager::beginTransaction()
