+++ lib/Doctrine/DBAL/TransactionDefinition.php
@@ -4,74 +4,37 @@ namespace Doctrine\DBAL;
 
 /**
  * Defines the parameters of a transaction.
+ *
+ * This object is immutable.
  */
 class TransactionDefinition
 {
     /**
-     * The transaction manager that created this definition.
-     *
-     * @var \Doctrine\DBAL\TransactionManager
-     */
-    private $transactionManager;
-
-    /**
-     * The isolation level for this transaction.
+     * The isolation level for the transaction, or null if not set.
      *
      * @var integer|null
      */
-    private $isolationLevel = null;
+    private $isolationLevel;
 
     /**
      * Class constructor.
      *
-     * @param TransactionManager $manager
+     * @param TransactionBuilder $builder A builder to get the parameters from, or null to use the defaults.
      */
-    public function __construct(TransactionManager $manager)
+    public function __construct(TransactionBuilder $builder = null)
     {
-        $this->transactionManager = $manager;
-    }
-
-    /**
-     * Sets the isolation level for this transaction.
-     *
-     * @param integer $isolationLevel One of the Connection::TRANSACTION_* constants.
-     *
-     * @return \Doctrine\DBAL\TransactionDefinition The current instance for chaining.
-     */
-    public function withIsolationLevel($isolationLevel)
-    {
-        $this->isolationLevel = $isolationLevel;
-
-        return $this;
+        if ($builder) {
+            $this->isolationLevel = $builder->getIsolationLevel();
+        }
     }
 
     /**
      * Returns the isolation level set for this transaction.
      *
-     * @return integer|null The isolation level if set, else null.
+     * @return integer|null The isolation level, or null if not set.
      */
     public function getIsolationLevel()
     {
         return $this->isolationLevel;
-    }
-
-    /**
-     * Begins the transaction and returns the associated Transaction object.
-     *
-     * @return \Doctrine\DBAL\Transaction
-     */
-    public function begin()
-    {
-        return $this->transactionManager->createTransaction($this);
-    }
-
-    /**
-     * Returns the transaction manager that created this definition.
-     *
-     * @return \Doctrine\DBAL\TransactionManager
-     */
-    public function getTransactionManager()
-    {
-        return $this->transactionManager;
     }
 }
